+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Please follow the format in [Keep a Changelog](http://keepachangelog.com/)
 
 ### Added
 
+- Support for ActiveRecord 4.2.x
+- Support for Mysql2 4.x
 - Allow passing any `pt-online-schema-change`'s arguments through the
    `PERCONA_ARGS` env var when executing a migration with `rake db:migrate:up`
    or `db:migrate:down`.